move configuration tests to their own file
